Rewriting Feature constructor
The aim of these modifications is to include the data structure of a
feature when a new instance is created. This way, even exisiting
features will be created with the corresponsing data structure.
This changes are made in order to create form groups where input types
are relatives to the data type.
Propeties are now an object with a type (string, boolean, date...) and a
value which are sent by the WFS server