Form - Managing missing value inside select boxes
Since the selected feature is updated whenever an input change, we had to force the user to change manipulate select boxes in order to trigger an update.
There is now an hint option that will ask for a value if undefined.
This is useful especially for required fields to avoid the server to throw an error.