App - Integrating new Header and Modal components
The new Header now makes part of the App component so it can be used to manage modals.
Modals are integrated to the App component but they are not rendered until their associated button is clicked.
For the moment, modals and their associated icons are defined withing the data attribute of the component.
A modal must have 4 attributes : `name`, `icon`, `visible` and `content`.